CleanOfferExtension
Clean Offer is a website for home (or in SF Apartment) buyers. The
script takes some of the tediousness out of the website by putting all
images in a quick view (no more scrolling). It consolidates a lot of
the page view information to make it easy to find open houses and other
information missing from easy viewing on the front page. You will need
a Clean Offer account to get started.